Extract helper class for GitHub to allow reuse in more CLI commands
