+++ developer-cli/Commands/BuildCommand.cs
