+++ developer-cli/Commands/CodeCleanupCommand.cs
