+++ developer-cli/Commands/CodeInspectionsCommand.cs
