+++ developer-cli/Commands/ConfigureContinuousDeploymentsCommand.cs
@@ -41,9 +41,7 @@ public class ConfigureContinuousDeploymentsCommand : Command
 
         PrintHeader("Collecting data");
 
-        SetGithubInfo();
-
-        LoginToGithub();
+        SetupToGithub();
 
         PublishExistingGithubVariables();
 
@@ -117,43 +115,10 @@ public class ConfigureContinuousDeploymentsCommand : Command
         AnsiConsole.WriteLine();
     }
 
-    private void SetGithubInfo()
+    private void SetupToGithub()
     {
-        // Get all Git remotes
-        var output = ProcessHelper.StartProcess("git remote -v", Configuration.GetSourceCodeFolder(), true);
+        Config.SetGithubInfo();
 
-        // Sort the output lines so that the "origin" is at the top
-        output = string.Join('\n', output.Split('\n').OrderBy(line => line.Contains("origin") ? 0 : 1));
-
-        var regex = new Regex(@"(?<githubUri>(https://github\.com/.*/.*\.git)|(git@github\.com:.*/.*\.git)) \(push\)");
-        var matches = regex.Matches(output);
-
-        var gitRemoteMatches = matches.Select(m => m.Groups["githubUri"].Value).ToArray();
-
-        var githubUri = string.Empty;
-        switch (gitRemoteMatches.Length)
-        {
-            case 0:
-                AnsiConsole.MarkupLine("[red]ERROR: No GitHub remote found. Please ensure you are within a Git repository with a GitHub.com as remote origin.[/]");
-                Environment.Exit(0);
-                break;
-            case 1:
-                githubUri = gitRemoteMatches.Single();
-                break;
-            case > 1:
-                githubUri = AnsiConsole.Prompt(new SelectionPrompt<string>()
-                    .Title("Select the GitHub remote")
-                    .AddChoices(gitRemoteMatches)
-                );
-                ProcessHelper.StartProcess($"gh repo set-default {githubUri}");
-                break;
-        }
-
-        Config.InitializeFromUri(githubUri);
-    }
-
-    private void LoginToGithub()
-    {
         if (!Config.IsLoggedIn())
         {
             ProcessHelper.StartProcess("gh auth login --git-protocol https --web");
@@ -915,24 +880,10 @@ public class Config
 
     public Dictionary<string, string> GithubVariables { get; set; } = new();
 
-    public void InitializeFromUri(string gitUri)
+    public void SetGithubInfo()
     {
-        string remote;
-        if (gitUri.StartsWith("https://github.com/"))
-        {
-            remote = gitUri.Replace("https://github.com/", "").Replace(".git", "");
-        }
-        else if (gitUri.StartsWith("git@github.com:"))
-        {
-            remote = gitUri.Replace("git@github.com:", "").Replace(".git", "");
-        }
-        else
-        {
-            throw new ArgumentException($"Invalid Git URI: {gitUri}. Only https:// and git@ formatted is supported.", nameof(gitUri));
-        }
-
-        var parts = remote.Split("/");
-        GithubInfo = new GithubInfo(parts[0], parts[1]);
+        var githubUri = GithubHelper.GetGithubUri();
+        GithubInfo = GithubHelper.GetGithubInfo(githubUri);
     }
 
     public bool IsLoggedIn()
@@ -941,17 +892,6 @@ public class Config
 
         return githubAuthStatus.Contains("Logged in to github.com");
     }
-}
-
-public class GithubInfo(string organizationName, string repositoryName)
-{
-    public string OrganizationName { get; } = organizationName;
-
-    public string RepositoryName { get; } = repositoryName;
-
-    public string Path => $"{OrganizationName}/{RepositoryName}";
-
-    public string Url => $"https://github.com/{Path}";
 }
 
 public record AzureSubscription(string Id, string Name, string TenantId, string State);
