+++ developer-cli/Commands/InstallCommand.cs
