+++ developer-cli/Commands/TranslateCommand.cs
