+++ developer-cli/Commands/UninstallCommand.cs
