+++ developer-cli/Installation/ChangeDetection.cs
