+++ developer-cli/Utilities/SolutionHelper.cs
